requirements.txt
biopython==1.79
deeplift==0.6.13.0
h5py==2.10.0
imbalanced_learn==0.8.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
pyfaidx==0.7.2.1
pyranges==0.0.120
scikit_learn==1.3.0
scipy==1.5.3
seaborn==0.11.2
tensorflow==1.14.0
tqdm==4.64.1
modisco==0.5.16.0
cudatoolkit==10.1.243
cudnn==7.6.5